Fix one-way unification and checking inferred types against explicit types
When one-way unifying an inferred type against a given type, the given
type should be instantiated at a higher level than the inferred type so
that we won't link type variables in it to other type variables in
itself (also fixed in this commit by not linking vars with same levels)
or type variables in the inferred type.

Also fix one-way unification argument order when comparing inferred and
given.

(Fixing these bugs reveals #5)

Fixes #4.