Actually fix one-way unification
The levels are to be able to choose a representative in union-sets and
easily prune level of a whole set.

Revert e5ccb0d, fix one-way unification by passing around a set of fixed
type variables for the right-hand side type.

With this fix we can finally implement `instance Functor []`.