Remove renaming pass
The pass was not that useful: we're not doing code generation and
optimizations so we don't move code around, and making each id unique is
not that useful (no shadowing, no accidental captures as we move code
etc.).

However it's making incremental updates more difficult as there's one
extra pass to incrementally run.

So at least for now we remove id.

`Id` type now becomes `String`. It's not ideal, but it works for now.